requirements.txt
capstone>=5.0.0
keystone-engine>=0.9.2
yara-python>=4.3.0
unicorn>=2.0.0
